+++ src/vue/App.vue
@@ -118,7 +118,8 @@ export default {
   methods: {
     ...mapActions({
       loadKvEntries: vuexTypes.LOAD_KV_ENTRIES,
-      loadDefaultQuoteAsset: vuexTypes.LOAD_DEFAULT_QUOTE_ASSET,
+      loadAssets: vuexTypes.LOAD_ASSETS,
+      loadAccount: vuexTypes.LOAD_ACCOUNT,
     }),
     async initApp () {
       api.useBaseURL(config.HORIZON_SERVER)
@@ -128,7 +129,6 @@ export default {
       api.useNetworkDetails(networkDetails)
 
       await this.loadKvEntries()
-      await this.loadDefaultQuoteAsset()
 
       if (this[vuexTypes.isLoggedIn]) {
         const wallet = new Wallet(
@@ -142,10 +142,12 @@ export default {
           'accountId': this[vuexTypes.walletAccountId],
           'email': this[vuexTypes.walletEmail],
         })
+        await this.loadAccount(this.walletAccountId)
       }
       walletsManager.useApi(api)
       factorsManager.useApi(api)
       documentsManager.useApi(api)
+      await this.loadAssets()
     },
     detectUnсompatibleBrowser () {
       this.isNotSupportedBrowser = !isCompatibleBrowser()