+++ src/vue/modules/withdrawal-fiat-card/index.vue
@@ -174,7 +174,7 @@ import { base } from '@tokend/js-sdk'
 import { api } from '@/api'
 import {
   required,
-  noMoreThanAvailableOnBalance,
+  lessThenMax,
   maxDecimalDigitsCount,
   cardNumber,
 } from '@validators'
@@ -246,7 +246,7 @@ export default {
         asset: { required },
         amount: {
           required,
-          noMoreThanAvailableOnBalance: noMoreThanAvailableOnBalance(
+          noMoreThanAvailableOnBalance: lessThenMax(
             this.form.asset.balance.value
           ),
           maxDecimalDigitsCount: maxDecimalDigitsCount(